requirements.txt
duckdb
polars
pyarrow
psycopg2-binary
pyyaml
python-dotenv
openpyxl
